requirements.txt
streamlit==1.26.0
openai==1.31.0
audio-recorder-streamlit==0.0.3
streamlit-float==0.3.5
python-dotenv==1.0.0
